make package installable with new minor releases
the language pack can currently not be installed with 2.1-rc01. This is because the minor version of the framework is also pushed, but the language pack still requires 100.0.  The '^' operator is designed for semantic versioning (which is used by the Magento Team) and should mostly take care, that your package is only installed in compatible magento instances (where Major Version = 100)